chore: move `findChain` to avoid issue `page without a React Component as default export`
